requirements.txt
pdfminer.six
chardet==3.0.4
nltk
fpdf
numpy
glob2
sklearn
matplotlib
pandas
toolz